Former-commit-id: 8f5e9fb3ad9f8a60605e7f9b9e087d97860ef6ce

+++ rec-common/src/main/java/es/capgemini/pfs/acuerdo/dto/DtoAcuerdo.java
@@ -28,7 +28,7 @@ public class DtoAcuerdo extends WebDto {
 	private Long periodo;
 	private String fechaCierre;
 	private String fechaLimite;
-	private Long importeCostas;	
+	private Float importeCostas;	
 	private String guid;
 	private Long idJBPM;
 	private Date fechaEstado;
@@ -217,12 +217,12 @@ public class DtoAcuerdo extends WebDto {
 		this.fechaLimite = fechaLimite;
 	}
 
-	public Long getImporteCostas() {
+	public Float getImporteCostas() {
 		return importeCostas;
 	}
 
 
-	public void setImporteCostas(Long importeCostas) {
+	public void setImporteCostas(Float importeCostas) {
 		this.importeCostas = importeCostas;
 	}
